Stake entity has contract and amount fields
modifys schema to include contract and amount fields for Stake entity. contract is address of the smart contract and is derived from call.inputs.scAddress value from the addStake() call. Amount is the amount of NXM tokens staked against the contract and is derived from call.inputs._amount from addStake()